+++ src/main/java/com/jakpop/stepsdictionary/data/service/CrudService.java
@@ -1,5 +1,6 @@
 package com.jakpop.stepsdictionary.data.service;
 
+import java.util.List;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -13,6 +14,14 @@ public abstract class CrudService<T, ID> {
 
     public Optional<T> get(ID id) {
         return this.getRepository().findById(id);
+    }
+
+    public List<T> getAll() {
+        return this.getRepository().findAll();
+    }
+
+    public T add(T entity) {
+        return this.getRepository().insert(entity);
     }
 
     public T update(T entity) {
